+++ CHANGES.md
@@ -7,6 +7,7 @@
 - MongoDB: Improve error reporting
 - MongoDB Full: Polars' `read_ndjson` doesn't load MongoDB JSON data well,
   use `fsspec` and `orjson` instead
+- MongoDB Full: Improved initialization of transformation subsystem
 
 ## 2024/09/19 v0.0.24
 - MongoDB Full: Refactor transformation subsystem to `commons-codec`